Handle missing vaadin CDI contexts on session expiration
Happens on TomEE, when it expires sessions on undeploy
- for example because of disabled session persistence.

See: http://tomee-openejb.979440.n4.nabble.com/session-expiration-upon-undeploy-vs-cdi-td4681169.html

+++ vaadin-cdi/src/main/java/com/vaadin/cdi/server/VaadinCDIServletService.java
@@ -44,6 +44,15 @@ public class VaadinCDIServletService extends VaadinServletService {
         @Override
         public void sessionDestroy(SessionDestroyEvent event) {
             getLogger().fine("VaadinSessionScopedContext destroy");
+            if (VaadinSessionScopedContext.guessContextIsUndeployed()) {
+                // Happens on tomcat when it expires sessions upon undeploy.
+                // beanManager.getPassivationCapableBean returns null for passivation id,
+                // so we would get an NPE from AbstractContext.destroyAllActive
+                getLogger().warning("VaadinSessionScoped context does not exist. " +
+                                "Maybe application is undeployed." +
+                                " Can't destroy VaadinSessionScopedContext.");
+                return;
+            }
             VaadinSessionScopedContext.destroy(event.getSession());
         }
